requirements.txt
PyQt6
robotframework
robotframework-pabot
pytest